Rename `rustc_resolve/errors.rs` into `rustc_resolve/error_helper.rs`
